+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Features
   1. [#1112](https://github.com/influxdata/chronograf/pull/1112): Add ability to delete a dashboard
   2. [#1120](https://github.com/influxdata/chronograf/pull/1120): Allow users to update user passwords.
+  3. [#1129](https://github.com/influxdata/chronograf/pull/1129): Allow InfluxDB and Kapacitor configuration via ENV vars or CLI options
 
 ### UI Improvements
   1. [#1101](https://github.com/influxdata/chronograf/pull/1101): Compress InfluxQL responses with gzip